requirements.txt
opencv_python==4.8.1.78
opencv_python_headless==4.8.1.78
pylsd_nova==1.2.1
tensorflow==2.13.0
gradio==3.50.2
gradio-client==0.6.1   

